+++ Source/PropertyTools.Wpf/PropertyGrid/PropertyGrid.cs
@@ -1022,6 +1022,7 @@ namespace PropertyTools.Wpf
                 else if (instance is INotifyDataErrorInfo notifyDataErrorInfoInstance)
                 {
                     tab.UpdateHasErrors(notifyDataErrorInfoInstance);
+                    //tab.UpdateTabForValidationResults(notifyDataErrorInfoInstance);
                 }
 
                 if (fillTab)
@@ -1370,7 +1371,8 @@ namespace PropertyTools.Wpf
                         propertyControl.Style = this.ValidationErrorStyle;
                     }
 
-                    errorControl = CreateErrorControl(pi, instance, tab);
+                    PropertyControlFactoryOptions options = new PropertyControlFactoryOptions { ValidationErrorTemplate = this.ValidationErrorTemplate };                    
+                    errorControl = this.ControlFactory.CreateErrorControl(pi, instance, tab, options);
 
                     // Add a row with the error control to the panel
                     // The error control is placed in column 1
@@ -1618,71 +1620,7 @@ namespace PropertyTools.Wpf
                     new Binding(pi.RadioDescriptor.Name) { Converter = new EnumToBooleanConverter() { EnumType = pi.RadioDescriptor.PropertyType }, ConverterParameter = pi.RadioValue });
             }
         }
-
-        /// <summary>
-        /// Creates the error control.
-        /// </summary>
-        private ContentControl CreateErrorControl(PropertyItem pi, object instance, Tab tab)
-        {
-            var dataErrorInfoInstance = instance as IDataErrorInfo;
-            var notifyDataErrorInfoInstance = instance as INotifyDataErrorInfo;
-
-            var errorControl = new ContentControl
-            {
-                ContentTemplate = this.ValidationErrorTemplate,
-                Focusable = false
-            };
-            IValueConverter errorConverter;
-            string propertyPath;
-            object source = null;
-            if (dataErrorInfoInstance != null)
-            {
-                errorConverter = new DataErrorInfoConverter(dataErrorInfoInstance, pi.PropertyName);
-                propertyPath = pi.PropertyName;
-                source = instance;
-            }
-            else
-            {
-                errorConverter = new NotifyDataErrorInfoConverter(notifyDataErrorInfoInstance, pi.PropertyName);
-                propertyPath = nameof(tab.HasErrors);
-                source = tab;
-                notifyDataErrorInfoInstance.ErrorsChanged += (s, e) =>
-                {
-                    tab.UpdateHasErrors(notifyDataErrorInfoInstance);
-                };
-            }
-
-            var visibilityBinding = new Binding(propertyPath)
-            {
-                Converter = errorConverter,
-                NotifyOnTargetUpdated = true,
-#if !NET40
-                ValidatesOnNotifyDataErrors = false,
-#endif
-                Source = source,
-            };
-
-            var contentBinding = new Binding(propertyPath)
-            {
-                Converter = errorConverter,
-#if !NET40
-                ValidatesOnNotifyDataErrors = false,
-#endif
-                Source = source,
-            };
-
-            errorControl.SetBinding(VisibilityProperty, visibilityBinding);
-
-            // When the visibility of the error control is changed, updated the HasErrors of the tab
-            errorControl.TargetUpdated += (s, e) =>
-            {
-                if (dataErrorInfoInstance != null)
-                    tab.UpdateHasErrors(dataErrorInfoInstance);
-            };
-            errorControl.SetBinding(ContentControl.ContentProperty, contentBinding);
-            return errorControl;
-        }
-
+        
         /// <summary>
         /// Creates the label control.
         /// </summary>